All changes until 2016/08/02
- RunAnalyzer -analyzeSpeed finished and tested.
- MathControll stringifyPaceFromSpeed:(float) bug fixed.
- Other calculation and math bugs fixed.
- RunAnalyzer -analyzeElevation -analyzeHeartRate finished. However, real device or Apple Watch is needed in order to test these methods.
- RunHelper -calculateSpeed:(NSMutableArray *) speed conversion bug fixed.
- RunHelper matrix (or similar) calculation implemented.
- Main.storyboard small changes.
- YQL warning fixed.
- Pod updated and modified.

Notes:
This is a very good commit. RunAnalyzer -analyzeSpeed is working. However, I still need to test out many more functions. RunAnalysis interface also need to be redesigned. ProgressVC is another important component of the smart run system (SRS). Based on goal, history (trend, improvements) user can see a week worth of training plan. Running is a difficult sport, thus allow users to set everyday reminders. Achievement will also help users to be motivated. Remember, Apple Watch Extension still need major improvements. I will create a note for specifics for the Apple Watch Extension. Good day!

+++ MoonRunner/Controller/ViewController/NewRunViewController.m
@@ -460,7 +460,7 @@ static NSString * const detailSegueName = @"NewRunDetails";
     NSLog(@"split %@", dict);
     time.text = [dict objectForKey:@"time"];
     distance.text = [dict objectForKey:@"distance"];
-    number.text = [NSString stringWithFormat:@"Split %li", indexPath.row + 1];
+    number.text = [NSString stringWithFormat:@"Split %i", (int)indexPath.row + 1];
     pace.text = [MathController stringifyAvgPaceFromDist:[[dict objectForKey:@"distance"] floatValue] overTime:[[dict objectForKey:@"time"] intValue]];
     return cell;
 }
